Создает модуляцию
